+++ Parser/Parser.cs
@@ -29,7 +29,7 @@ public class Parser
     {
         var index = 0;
         Scope scope = new(Logger);
-        while (index + 1 <= _tokens.Count())
+        while(index + 1 <= _tokens.Count())
         {
             if(_tokens == null) 
             {
@@ -44,6 +44,10 @@ public class Parser
             scope.AddNode(parseResult.node);
             index = parseResult.index;
             index++;
+            if (_tokens[index].TokenType.Equals(TokenSyntax.SEMICOLON))
+            {
+                index++;
+            }
         }
 
         Logger.Log("Finished Parsing", this.GetType().Name, LogType.INFO);